Fixed compilation issues
Apparently $(COMPILE.cc) and $(LINK.cc) correspond to C++, not CC like
the compiler. Replacing with .c is more appropriate, fixed this.

Additionally, forgot a comma after my "help" string, causing the typedef
for PetscErrorCode to not be valid, causing every line where that type
was used to throw an error. Gross.